refactor(zeroclaw-tools): extract duplicate domain/URL validation into shared helpers
Consolidate `normalize_allowed_domains`, `normalize_domain`,
`host_matches_allowlist`, and `is_private_or_local_host` from
`browser.rs`, `browser_open.rs`, `http_request.rs`, and `web_fetch.rs`
into a new `helpers::domain_guard` module. This reduces code
duplication and provides a single source of truth for domain-related
validation across all tools.
