+++ crates/zeroclaw-tools/src/helpers/mod.rs
@@ -1,0 +1,1 @@
+pub mod domain_guard;
