+++ crates/zeroclaw-tools/src/web_fetch.rs
@@ -1,3 +1,4 @@
+use crate::helpers::domain_guard;
 use async_trait::async_trait;
 use futures_util::StreamExt;
 use serde_json::json;
@@ -42,15 +43,15 @@ impl WebFetchTool {
     ) -> anyhow::Result<Self> {
         Ok(Self {
             security,
-            allowed_domains: normalize_allowed_domains(
+            allowed_domains: domain_guard::normalize_allowed_domains(
                 allowed_domains,
                 "web_fetch.allowed_domains",
             )?,
-            blocked_domains: normalize_allowed_domains(
+            blocked_domains: domain_guard::normalize_allowed_domains(
                 blocked_domains,
                 "web_fetch.blocked_domains",
             )?,
-            allowed_private_hosts: normalize_allowed_domains(
+            allowed_private_hosts: domain_guard::normalize_allowed_domains(
                 allowed_private_hosts,
                 "web_fetch.allowed_private_hosts",
             )?,
@@ -536,11 +537,11 @@ fn validate_target_url_with_dns_check(
     let host = extract_host(url)?;
 
     // blocked_domains always takes precedence
-    if host_matches_allowlist(&host, blocked_domains) {
+    if domain_guard::host_matches_allowlist(&host, blocked_domains) {
         anyhow::bail!("Host '{host}' is in {tool_name}.blocked_domains");
     }
 
-    let host_is_private_or_local = is_private_or_local_host(&host);
+    let host_is_private_or_local = domain_guard::is_private_or_local_host(&host);
     let private_host_allowed =
         host_matches_private_allowlist(&host, allowed_private_hosts, host_is_private_or_local);
 
@@ -561,7 +562,12 @@ fn validate_target_url_with_dns_check(
         );
     }
 
-    if !private_host_allowed && !host_matches_allowlist(&host, allowed_domains) {
+    // Private hosts in the allowlist skip the allowed_domains check.
+    // Non-private hosts still require allowed_domains approval even when
+    // listed in allowed_private_hosts (e.g. explicit internal DNS names).
+    let skip_allowed_domains = host_is_private_or_local && private_host_allowed;
+
+    if !skip_allowed_domains && !domain_guard::host_matches_allowlist(&host, allowed_domains) {
         anyhow::bail!("Host '{host}' is not in {tool_name}.allowed_domains");
     }
 
@@ -585,68 +591,6 @@ fn append_chunk_with_cap(buffer: &mut Vec<u8>, chunk: &[u8], hard_cap: usize) ->
 
     buffer.extend_from_slice(chunk);
     buffer.len() >= hard_cap
-}
-
-fn normalize_allowed_domains(domains: Vec<String>, label: &str) -> anyhow::Result<Vec<String>> {
-    let mut rejected = Vec::new();
-    let mut normalized = domains
-        .into_iter()
-        .filter_map(|d| {
-            normalize_domain(&d).or_else(|| {
-                rejected.push(d.clone());
-                None
-            })
-        })
-        .collect::<Vec<_>>();
-    if !rejected.is_empty() {
-        anyhow::bail!(
-            "Invalid {label} entry(s): [{}]. Each entry must be a valid domain, hostname, IPv4, or IPv6 address.",
-            rejected.join(", ")
-        );
-    }
-    normalized.sort_unstable();
-    normalized.dedup();
-    Ok(normalized)
-}
-
-fn normalize_domain(raw: &str) -> Option<String> {
-    let input = raw.trim();
-    if input.is_empty() || input.chars().any(char::is_whitespace) {
-        return None;
-    }
-
-    let bare_ip = match (input.starts_with('['), input.ends_with(']')) {
-        (true, true) => &input[1..input.len() - 1],
-        (false, false) => input,
-        _ => return None,
-    };
-    if let Ok(ip) = bare_ip.parse::<std::net::IpAddr>() {
-        return Some(ip.to_string().to_lowercase());
-    }
-
-    let parsed = reqwest::Url::parse(input)
-        .or_else(|_| reqwest::Url::parse(&format!("https://{input}")))
-        .ok()?;
-
-    if !parsed.username().is_empty() || parsed.password().is_some() {
-        return None;
-    }
-
-    let host = parsed.host_str()?;
-    let trimmed = host.trim();
-    let host_no_brackets = match (trimmed.starts_with('['), trimmed.ends_with(']')) {
-        (true, true) => &trimmed[1..trimmed.len() - 1],
-        (false, false) => trimmed,
-        _ => return None,
-    };
-    let normalized = host_no_brackets
-        .trim_start_matches('.')
-        .trim_end_matches('.');
-    if normalized.is_empty() {
-        return None;
-    }
-
-    Some(normalized.to_lowercase())
 }
 
 fn extract_host(url: &str) -> anyhow::Result<String> {
@@ -702,59 +646,15 @@ fn extract_host(url: &str) -> anyhow::Result<String> {
     Ok(host)
 }
 
-fn host_matches_allowlist(host: &str, allowed_domains: &[String]) -> bool {
-    if allowed_domains.iter().any(|domain| domain == "*") {
-        return true;
-    }
-
-    allowed_domains.iter().any(|domain| {
-        host == domain
-            || host
-                .strip_suffix(domain)
-                .is_some_and(|prefix| prefix.ends_with('.'))
-    })
-}
-
 fn host_matches_private_allowlist(
     host: &str,
     allowed_private_hosts: &[String],
     host_is_private_or_local: bool,
 ) -> bool {
-    allowed_private_hosts.iter().any(|domain| {
-        if domain == "*" {
-            host_is_private_or_local
-        } else {
-            host == domain
-                || host
-                    .strip_suffix(domain)
-                    .is_some_and(|prefix| prefix.ends_with('.'))
-        }
-    })
-}
-
-fn is_private_or_local_host(host: &str) -> bool {
-    let bare = host
-        .strip_prefix('[')
-        .and_then(|h| h.strip_suffix(']'))
-        .unwrap_or(host);
-
-    let has_local_tld = bare
-        .rsplit('.')
-        .next()
-        .is_some_and(|label| label == "local");
-
-    if bare == "localhost" || bare.ends_with(".localhost") || has_local_tld {
-        return true;
+    if allowed_private_hosts.iter().any(|d| d == "*") {
+        return host_is_private_or_local;
     }
-
-    if let Ok(ip) = bare.parse::<std::net::IpAddr>() {
-        return match ip {
-            std::net::IpAddr::V4(v4) => is_non_global_v4(v4),
-            std::net::IpAddr::V6(v6) => is_non_global_v6(v6),
-        };
-    }
-
-    false
+    domain_guard::host_matches_allowlist(host, allowed_private_hosts)
 }
 
 #[cfg(not(test))]
@@ -795,8 +695,8 @@ fn validate_resolved_ips_are_public(host: &str, ips: &[std::net::IpAddr]) -> any
 
     for ip in ips {
         let non_global = match ip {
-            std::net::IpAddr::V4(v4) => is_non_global_v4(*v4),
-            std::net::IpAddr::V6(v6) => is_non_global_v6(*v6),
+            std::net::IpAddr::V4(v4) => domain_guard::is_non_global_v4(*v4),
+            std::net::IpAddr::V6(v6) => domain_guard::is_non_global_v6(*v6),
         };
         if non_global {
             anyhow::bail!("Blocked host '{host}' resolved to non-global address {ip}");
@@ -804,33 +704,6 @@ fn validate_resolved_ips_are_public(host: &str, ips: &[std::net::IpAddr]) -> any
     }
 
     Ok(())
-}
-
-fn is_non_global_v4(v4: std::net::Ipv4Addr) -> bool {
-    let [a, b, c, _] = v4.octets();
-    v4.is_loopback()
-        || v4.is_private()
-        || v4.is_link_local()
-        || v4.is_unspecified()
-        || v4.is_broadcast()
-        || v4.is_multicast()
-        || (a == 100 && (64..=127).contains(&b))
-        || a >= 240
-        || (a == 192 && b == 0 && (c == 0 || c == 2))
-        || (a == 198 && b == 51)
-        || (a == 203 && b == 0)
-        || (a == 198 && (18..=19).contains(&b))
-}
-
-fn is_non_global_v6(v6: std::net::Ipv6Addr) -> bool {
-    let segs = v6.segments();
-    v6.is_loopback()
-        || v6.is_unspecified()
-        || v6.is_multicast()
-        || (segs[0] & 0xfe00) == 0xfc00
-        || (segs[0] & 0xffc0) == 0xfe80
-        || (segs[0] == 0x2001 && segs[1] == 0x0db8)
-        || v6.to_ipv4_mapped().is_some_and(is_non_global_v4)
 }
 
 #[cfg(test)]
@@ -1030,19 +903,6 @@ mod tests {
             .unwrap_err()
             .to_string();
         assert!(err.contains("local/private"));
-    }
-
-    #[test]
-    fn ssrf_blocks_loopback() {
-        assert!(is_private_or_local_host("127.0.0.1"));
-        assert!(is_private_or_local_host("127.0.0.2"));
-    }
-
-    #[test]
-    fn ssrf_blocks_rfc1918() {
-        assert!(is_private_or_local_host("10.0.0.1"));
-        assert!(is_private_or_local_host("172.16.0.1"));
-        assert!(is_private_or_local_host("192.168.1.1"));
     }
 
     #[test]
@@ -1257,43 +1117,6 @@ mod tests {
     }
 
     // ── Domain normalization ─────────────────────────────────────
-
-    #[test]
-    fn normalize_domain_strips_scheme_and_case() {
-        let got = normalize_domain("  HTTPS://Docs.Example.com/path ").unwrap();
-        assert_eq!(got, "docs.example.com");
-    }
-
-    #[test]
-    fn normalize_domain_rejects_userinfo() {
-        assert!(normalize_domain("https://user@example.com").is_none());
-        assert!(normalize_domain("user@example.com").is_none());
-        assert!(normalize_domain("https://user:pass@example.com").is_none());
-        assert!(normalize_domain("user:pass@example.com").is_none());
-    }
-
-    #[test]
-    fn normalize_domain_rejects_unmatched_brackets() {
-        assert!(normalize_domain("[::1").is_none());
-        assert!(normalize_domain("::1]").is_none());
-        assert!(normalize_domain("[127.0.0.1").is_none());
-        assert!(normalize_domain("127.0.0.1]").is_none());
-    }
-
-    #[test]
-    fn normalize_deduplicates() {
-        let got = normalize_allowed_domains(
-            vec![
-                "example.com".into(),
-                "EXAMPLE.COM".into(),
-                "https://example.com/".into(),
-            ],
-            "test",
-        )
-        .unwrap();
-        assert_eq!(got, vec!["example.com".to_string()]);
-    }
-
     // ── Blocked domains ──────────────────────────────────────────
 
     #[test]
@@ -1783,6 +1606,26 @@ mod tests {
             result.is_ok(),
             "allowlisted private domain was rejected: {result:?}"
         );
+    }
+
+    #[test]
+    fn private_allowlist_explicit_entry_must_pass_allowed_domains() {
+        let allowed_domains = vec!["example.com".to_string()];
+        let blocked_domains = vec![];
+        let allowed_private_hosts = vec!["unrelated.com".to_string()];
+
+        let err = validate_target_url_with_dns_check(
+            "https://unrelated.com/api",
+            &allowed_domains,
+            &blocked_domains,
+            &allowed_private_hosts,
+            "web_fetch",
+            |_| anyhow::Ok(()),
+        )
+        .unwrap_err()
+        .to_string();
+
+        assert!(err.contains("allowed_domains"), "unexpected error: {err}");
     }
 
     #[test]
